rename: Proposal.java와 Report.java를 domain 패키지 하위 dev 패키지로 이동. (#4)
