+++ src/main/java/com/newfit/reservation/domain/dev/Proposal.java
@@ -1,5 +1,6 @@
-package com.newfit.reservation.domain;
+package com.newfit.reservation.domain.dev;
 
+import com.newfit.reservation.domain.User;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -9,9 +10,9 @@ import lombok.*;
                                                     // 다만, Entity 객체를 직접 생성자로 생성할 일은 없을듯 하여 protected 로 설정했습니다.
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
-public class Report {
+public class Proposal {
 
-    // Report 테이블 PK 입니다.
+    // Proposal 테이블 PK 입니다.
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
@@ -20,18 +21,18 @@ public class Report {
     @JoinColumn(name = "user_id")
     private User user;
 
-    // 버그 제목을 나타냅니다.
+    // 기능 이름을 나타냅니다.
     @Column(nullable = false)
-    private String subject;
+    private String name;
 
-    // 버그 내용을 나타냅니다.
+    // 기능 내용을 나타냅니다.
     @Column(nullable = false)
     private String content;
 
     // 연관관계 편의 메소드입니다.
     public void setUserRelation(User user) {
         this.user = user;
-        user.getReportList().add(this);
+        user.getProposalList().add(this);
     }
 }
